fix(dev/mojo): Fix the wasm twins of three gui sites
dev/mojo/wasm has no check of its own, so building it by hand is the only
way to see it. Three sites that the gui copies already had fixed were still
broken here, all in files that differ from their gui counterparts and so
were not covered by them: the Int-to-UInt32 conversion in
context.handler_count and in bench's three modulo operands, and the unused
loop index in main.mojo.

web/src/main.mojo now compiles under the project's own -Werror with no
diagnostics at all.

+++ dev/mojo/wasm/core/src/component/context.mojo
@@ -2098,7 +2098,7 @@ struct ComponentContext(Movable):
 
         Useful for testing and introspection.
         """
-        return self.shell.runtime[0].handler_count()
+        return UInt32(self.shell.runtime[0].handler_count())
 
     fn view_events(self) -> List[EventBinding]:
         """Return a copy of the registered view event bindings.
